requirements.txt
click
openai
PyYAML
pytest